MY MAIN PROJECT

+++ RestapiSpring/src/main/java/com/RestProject1/springrestapi/service/BasicEmployeeService.java
@@ -7,6 +7,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.RestProject1.springrestapi.Converter.Convertor;
 import com.RestProject1.springrestapi.DTO.EmployeeDTO;
 import com.RestProject1.springrestapi.model.Employee;
 import com.RestProject1.springrestapi.repository.BenifitsRepository;
@@ -22,20 +23,32 @@ public class BasicEmployeeService implements EmployeeService{
 	
 	 @Autowired
 	 private ModelMapper modelMapper;
+	 
+	 @Autowired
+	 private Convertor convert;
+	 
 
 	
 	@Override
 	public List<Employee> getEmployees() {
+		
+		//List<Employee> emp = empRepository.findAll();
+		//List<EmployeeDTO> dto = convert.entityToDto(emp);
 		return empRepository.findAll();
 	}
 
 
 
 	@Override
-	public Employee saveEmployee(Employee e) {
+	public Employee saveEmployee(Employee emp) {
 		
+		/*Employee e = new Employee();
+		EmployeeDTO d = new EmployeeDTO();
 		
-		return empRepository.save(e);
+		e = convert.dtoToEntity(dto);
+		empRepository.save(e);
+		d = convert.entityToDto(e);*/
+		return empRepository.save(emp);
 	}
 
 
@@ -44,6 +57,7 @@ public class BasicEmployeeService implements EmployeeService{
 	public Employee getEmployee(Long id) {
 		
 		Optional<Employee> emp =  empRepository.findById(id);
+				
 		if(emp.isPresent())
 		{
 			return emp.get();
@@ -56,7 +70,14 @@ public class BasicEmployeeService implements EmployeeService{
 
 	@Override
 	public void deleteEmployee(Long id) {
-		empRepository.deleteById(id);
+		
+		Optional<Employee> emp =  empRepository.findById(id);
+		if(emp.isPresent())
+		{
+			empRepository.deleteById(id);
+		}
+		else
+			throw new RuntimeException("Employee is not found for the ID : " + id);
 		
 	}
 
@@ -77,7 +98,7 @@ public class BasicEmployeeService implements EmployeeService{
 	}
 	
 	
-	public Employee dtoToEmp(EmployeeDTO dto)
+/*	public Employee dtoToEmp(EmployeeDTO dto)
 	{
 		Employee emp = this.modelMapper.map(dto, Employee.class);
 		return emp;
@@ -87,6 +108,6 @@ public class BasicEmployeeService implements EmployeeService{
 	{
 		EmployeeDTO dto = this.modelMapper.map(e, EmployeeDTO.class);
 		return dto;
-	}
+	} */
 
 }
